Delete libtock stuff we don't need; add no-op impl for the host so that the compiles are clean
