+++ Cargo.toml
@@ -18,8 +18,34 @@ members = [
     "rom",
     "romtime",
     "runtime",
-    "runtime/capsules",
     "runtime/apps/pldm",
+    "runtime/apps/libtock/apis/interface/buttons",
+    "runtime/apps/libtock/apis/interface/buzzer",
+    "runtime/apps/libtock/apis/interface/console",
+    "runtime/apps/libtock/apis/interface/leds",
+    "runtime/apps/libtock/apis/kernel/low_level_debug",
+    "runtime/apps/libtock/apis/peripherals/adc",
+    "runtime/apps/libtock/apis/peripherals/alarm",
+    "runtime/apps/libtock/apis/peripherals/gpio",
+    "runtime/apps/libtock/apis/peripherals/i2c_master",
+    "runtime/apps/libtock/apis/peripherals/i2c_master_slave",
+    "runtime/apps/libtock/apis/peripherals/rng",
+    "runtime/apps/libtock/apis/sensors/air_quality",
+    "runtime/apps/libtock/apis/sensors/ambient_light",
+    "runtime/apps/libtock/apis/sensors/ninedof",
+    "runtime/apps/libtock/apis/sensors/proximity",
+    "runtime/apps/libtock/apis/sensors/temperature",
+    "runtime/apps/libtock/apis/storage/key_value",
+    "runtime/apps/libtock/panic_handlers/debug_panic",
+    "runtime/apps/libtock/panic_handlers/small_panic",
+    "runtime/apps/libtock/platform",
+    "runtime/apps/libtock/runner",
+    "runtime/apps/libtock/runtime",
+    "runtime/apps/libtock/syscalls_tests",
+    "runtime/apps/libtock/tools/print_sizes",
+    "runtime/apps/libtock/ufmt",
+    "runtime/apps/libtock/unittest",
+    "runtime/capsules",
     "runtime/i3c",
     "tests/hello",
     "tests/integration",
