+++ runtime/apps/libtock/apis/interface/buttons/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
 license = "Apache-2.0 OR MIT"
 edition = "2018"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 description = "libtock buttons driver"
 
 [dependencies]
