+++ runtime/apps/libtock/apis/net/ieee802154/Cargo.toml
@@ -1,13 +1,10 @@
 [package]
 name = "libtock_ieee802154"
 version = "0.1.0"
-authors = [
-    "Tock Project Developers <tock-dev@googlegroups.com>",
-]
+authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
 license = "Apache-2.0 OR MIT"
 edition = "2021"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 description = "libtock raw IEEE 802.15.4 stack driver"
 
 [dependencies]
