+++ runtime/apps/libtock/apis/peripherals/alarm/Cargo.toml
@@ -8,7 +8,6 @@ authors = [
 license = "Apache-2.0 OR MIT"
 edition = "2018"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 description = "libtock alarm driver"
 
 [dependencies]
