+++ runtime/apps/libtock/apis/peripherals/i2c_master_slave/Cargo.toml
@@ -8,9 +8,7 @@ authors = [
 license = "Apache-2.0 OR MIT"
 edition = "2021"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 description = "libtock I2C master-slave driver"
 
 [dependencies]
 libtock_platform = { path = "../../../platform" }
-
