+++ runtime/apps/libtock/apis/peripherals/rng/Cargo.toml
@@ -8,7 +8,6 @@ authors = [
 license = "Apache-2.0 OR MIT"
 edition = "2021"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 description = "libtock rng driver"
 
 [dependencies]
