+++ runtime/apps/libtock/apis/sensors/proximity/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
 license = "Apache-2.0 OR MIT"
 edition = "2021"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 description = "libtock proximity driver"
 
 [dependencies]
@@ -13,4 +12,3 @@ libtock_platform = { path = "../../../platform" }
 
 [dev-dependencies]
 libtock_unittest = { path = "../../../unittest" }
-
