+++ runtime/apps/libtock/apis/storage/key_value/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
 name = "libtock_key_value"
 version = "0.1.0"
-authors = [
-    "Tock Project Developers <tock-dev@googlegroups.com>",
-]
+authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
 license = "MIT/Apache-2.0"
 edition = "2021"
 repository = "https://www.github.com/tock/libtock-rs"
