--- runtime/apps/libtock/demos/st7789-slint/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "st7789-slint"
-version = "0.1.0"
-edition = "2021"
-rust-version.workspace = true
-authors = ["Alistair Francis <alistair.francis@wdc.com>"]
-description = """A demo to use the Slint GUI library with a ST7789 display via SPI using libtock-rs."""
-license = "Apache-2.0 OR MIT"
-
-[dependencies]
-libtock = { path = "../../", features = ["rust_embedded"] }
-
-embedded-hal = "1.0"
-
-mipidsi = "0.8.0"
-display-interface-spi = "0.5"
-embedded-graphics = "0.8"
-
-# The heap allocator and portable atomics
-embedded-alloc = "0.5.1"
-critical-section = "1.0"
-
-slint = { git = "https://github.com/slint-ui/slint", default-features = false, features = ["libm", "unsafe-single-threaded"] }
-mcu-board-support = { git = "https://github.com/slint-ui/slint" }
-
-display-interface = "0.5"
-
-[build-dependencies]
-libtock_build_scripts = { path = "../../build_scripts" }
-
-slint-build = { git = "https://github.com/slint-ui/slint" }
