--- runtime/apps/libtock/demos/st7789-slint/ui/appwindow.slint
@@ -1,9 +1,0 @@
-import {Button, AboutSlint} from "std-widgets.slint";
-
-export component MainWindow inherits Window {
-    width: 240px;
-    height: 240px;
-    VerticalLayout {
-        AboutSlint { }
-    }
-}
