--- runtime/apps/libtock/demos/st7789/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    libtock_build_scripts::auto_layout();
-}
