+++ runtime/apps/libtock/panic_handlers/small_panic/Cargo.toml
@@ -6,7 +6,6 @@ description = """Small and basic panic handler for libtock. Calls low-level debu
 edition = "2021"
 license = "Apache-2.0 OR MIT"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 
 [dependencies]
 libtock_low_level_debug = { path = "../../apis/kernel/low_level_debug" }
