+++ runtime/apps/libtock/platform/Cargo.toml
@@ -10,7 +10,6 @@ edition = "2021"
 license = "Apache-2.0 OR MIT"
 name = "libtock_platform"
 repository = "https://www.github.com/tock/libtock/rs"
-rust-version.workspace = true
 version = "0.1.0"
 
 [features]
