+++ runtime/apps/libtock/runner/Cargo.toml
@@ -6,7 +6,6 @@ license = "Apache-2.0 OR MIT"
 name = "runner"
 publish = false
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 version = "0.1.0"
 
 [dependencies]
