+++ runtime/apps/libtock/tools/print_sizes/Cargo.toml
@@ -7,7 +7,6 @@
 authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
 edition = "2018"
 name = "print_sizes"
-rust-version.workspace = true
 version = "0.1.0"
 
 [dependencies]
