+++ runtime/apps/libtock/unittest/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 license = "Apache-2.0 OR MIT"
 name = "libtock_unittest"
 repository = "https://www.github.com/tock/libtock-rs"
-rust-version.workspace = true
 version = "0.1.0"
 
 [dependencies]
